--- Source/GameBaseFramework/Private/GBFSaveGame.cpp
@@ -1,1 +1,0 @@
-#include "GBFSaveGame.h"
